VMO-6650 extract expression-generating helper and add unit tests
