+++ src/components/interaction-designer/block-editors/choices/index.ts
@@ -1,0 +1,1 @@
+export * from './SelectOneResponseBlockContactPropertyEditor.vue'
